+++ Src/MoneyFox.Presentation/ViewModels/ModifyPaymentViewModel.cs
@@ -43,8 +43,8 @@ namespace MoneyFox.Presentation.ViewModels
         PaymentViewModel SelectedPayment { get; }
 
         /// <summary>
-        ///     Property to format amount string to double with the proper culture.
-        ///     This is used to prevent issues when converting the amount string to double
+        ///     Property to format amount string to decimal with the proper culture.
+        ///     This is used to prevent issues when converting the amount string to decimal
         ///     without the correct culture.
         /// </summary>
         string AmountString { get; }
@@ -211,8 +211,8 @@ namespace MoneyFox.Presentation.ViewModels
         }
 
         /// <summary>
-        ///     Property to format amount string to double with the proper culture.
-        ///     This is used to prevent issues when converting the amount string to double
+        ///     Property to format amount string to decimal with the proper culture.
+        ///     This is used to prevent issues when converting the amount string to decimal
         ///     without the correct culture.
         /// </summary>
         public string AmountString
@@ -222,7 +222,7 @@ namespace MoneyFox.Presentation.ViewModels
             {
                 // we remove all separator chars to ensure that it works in all regions
                 string amountString = HelperFunctions.RemoveGroupingSeparators(value);
-                if (double.TryParse(amountString, NumberStyles.Any, CultureInfo.CurrentCulture, out double convertedValue))
+                if (decimal.TryParse(amountString, NumberStyles.Any, CultureInfo.CurrentCulture, out decimal convertedValue))
                 {
                     SelectedPayment.Amount = convertedValue;
                 }
